+++ src/client/app/+system-users/system-users.component.ts
@@ -1,7 +1,7 @@
 import { Component,OnInit,ViewChild } from '@angular/core';
 import {HTTP_PROVIDERS} from '@angular/http';
 import {HttpService} from '../services/http-service';
-import { DynamicCmp} from '../shared/dialog/dialog';
+import { DialogComponent} from '../shared/dialog/dialog.component';
 import { ValidationService } from '../shared/dialog/validation.service';
 import { ControlMessagesComponent } from '../shared/dialog/control-messages.component';
 import { FormBuilder, FormGroup, Validators,FormControl,REACTIVE_FORM_DIRECTIVES } from '@angular/forms';
@@ -12,14 +12,15 @@ import { FormBuilder, FormGroup, Validators,FormControl,REACTIVE_FORM_DIRECTIVES
   moduleId: module.id,
   selector: 'ap-system-users',
    providers: [HttpService,HTTP_PROVIDERS],
-   directives:[DynamicCmp],
+   directives:[DialogComponent],
   templateUrl: 'system-users.component.html',
  
 })
 export class SystemUsersComponent implements OnInit {
-    @ViewChild(DynamicCmp) modalSystemUsersComponent: DynamicCmp;
+  
    itemsObservables:any;
-systemUserData:any;
+systemUserData:Array<any>;
+  @ViewChild(DialogComponent) modalSystemUsersComponent: DialogComponent;
 userForm:any;
 systemUsersformData:any;
 systemUsersForm:any;
