+++ src/client/app/shared/data.json
@@ -1,0 +1,1 @@
+{}
