+++ src/client/app/shared/dialog/control-messages.component.ts
@@ -14,9 +14,8 @@ export class ControlMessagesComponent {
     for (let propertyName in this.control.errors) {
       if (this.control.errors.hasOwnProperty(propertyName) && this.control.touched) {
         return ValidationService.getValidatorErrorMessage(propertyName, this.control.errors[propertyName]);
-      }
+      }   
     }
-    
     return null;
   }
 }
