+++ src/client/app/shared/dialog/index.ts
@@ -1,1 +1,1 @@
-export * from './dialog';
+export * from './dialog.component';
