+++ src/client/index.html
@@ -27,7 +27,7 @@
 
     <!-- Modernizr -->
     <script src="assets/js/modernizr-2.8.3.min.js"></script>
-    <link href='http://fonts.googleapis.com/css?family=Source+Sans+Pro:400,700' rel='stylesheet' type='text/css'>
+    <!--<link href='http://fonts.googleapis.com/css?family=Source+Sans+Pro:400,700' rel='stylesheet' type='text/css'>-->
 
     <!-- JS -->
     <script src="assets/js/jquery-1.11.3.min.js"></script>
